requirements.txt
six
tqdm
torch==1.3.1
torchtext==0.4.0
future
configargparse
spacy==2.2.4
scikit-learn==0.22.1
pandas==1.0.5
numpy==1.17.4
pytorch-pretrained-bert==0.6.2
py-rouge==1.1
nltk==3.4.5
pyyaml
